Task #13

+++ src/main/java/pl/futurecollars/invoicing/controller/invoice/InvoiceApi.java
@@ -23,7 +23,7 @@ public interface InvoiceApi {
 
   @ApiOperation(value = "Add new invoice to system")
   @PostMapping
-  int add(@RequestBody Invoice invoice);
+  long add(@RequestBody Invoice invoice);
 
   @ApiOperation(value = "Get invoice by id")
   @GetMapping(value = "/{id}")
